% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_diag_hatvalues.R
\name{block_diag_hatvalues}
\alias{block_diag_hatvalues}
\title{Block Diagonal Hat Matrix}
\usage{
block_diag_hatvalues(model, cl = NULL)
}
\arguments{
\item{model}{A \code{fixest} model object.}

\item{cl}{A vector of cluster membership with the same length as the number
of observations used in \code{model}.}
}
\value{
A \code{nobs} by \code{nobs} sparse matrix of the block diagonal hat matrix.
}
\description{
Calculates the block diagonal hat matrix for a fixest model where the blocks
correspond to clusters (cl_i == cl_j) otherwise the (i,j) element is
zeroed out. Note the data doesn't have to be sorted for this.
}
