% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{coef.jive_est}
\alias{coef.jive_est}
\title{Extract coefficient from a \code{jive_est} object}
\usage{
\method{coef}{jive_est}(x)
}
\arguments{
\item{x}{A \code{jive_est} object from \code{jive}/\code{ujive}/\code{ijive}}
}
\description{
Extract coefficient from a \code{jive_est} object
}
