% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fml_breaker}
\alias{fml_breaker}
\title{Break apart formula (from right to left) based on a symbole (\code{~} or \code{|})}
\usage{
fml_breaker(fml, op)
}
\arguments{
\item{fml}{Formula following \code{fixest} syntax.}

\item{op}{String. Either \code{~} or \code{|}}
}
\value{
list of \code{symbol} or \code{language} from right to left that are split at each occurence of \code{op}.
}
\description{
Break apart formula (from right to left) based on a symbole (\code{~} or \code{|})
}
