% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_diag_utils.R
\name{mult_DX_T}
\alias{mult_DX_T}
\title{Computes D_X \%*\% T for diagonal block projection matrix}
\usage{
mult_DX_T(model, T, cl = NULL)
}
\arguments{
\item{model}{A \code{fixest} model object.}

\item{T}{Numeric vector}

\item{cl}{A vector of cluster membership with the same length as the number
of observations used in \code{model}.}
}
\description{
Computes D_X \%*\% T for diagonal block projection matrix
}
