% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.manyiv_est}
\alias{print.manyiv_est}
\title{A print method for \code{manyiv_est} objects}
\usage{
\method{print}{manyiv_est}(x, ...)
}
\arguments{
\item{x}{A \code{manyiv_est} object from \code{manyiv}}

\item{...}{Unused}
}
\description{
A print method for \code{manyiv_est} objects
}
