% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{se.jive_est}
\alias{se.jive_est}
\title{Extract standard error from a \code{jive_est} object}
\usage{
se.jive_est(x)
}
\arguments{
\item{x}{A \code{jive_est} object from \code{jive}/\code{ujive}/\code{ijive}}
}
\description{
Extract standard error from a \code{jive_est} object
}
