% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_diag_utils.R
\name{solve_ImDX1pDX2_T}
\alias{solve_ImDX1pDX2_T}
\title{Computes (I - (D_X1 - D_X2))^{-1} T for diagonal block projection matrix}
\usage{
solve_ImDX1pDX2_T(model1, model2, T, cl = NULL)
}
\arguments{
\item{model1}{A \code{fixest} model object. This should be the complete model (W Z)}

\item{model2}{A \code{fixest} model object. This should be the simpler model (W)}

\item{T}{Numeric vector}

\item{cl}{A vector of cluster membership with the same length as the number
of observations used in \code{model}.}
}
\description{
Computes (I - (D_X1 - D_X2))^{-1} T for diagonal block projection matrix
}
