% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_diag_utils.R
\name{solve_ImDX_T}
\alias{solve_ImDX_T}
\title{Computes (I - D_X)^{-1} T for diagonal block projection matrix}
\usage{
solve_ImDX_T(model, T, cl = NULL)
}
\arguments{
\item{model}{A \code{fixest} model object.}

\item{T}{Numeric vector}

\item{cl}{A vector of cluster membership with the same length as the number
of observations used in \code{model}.}
}
\description{
Computes (I - D_X)^{-1} T for diagonal block projection matrix
}
