% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.manyiv_est}
\alias{summary.manyiv_est}
\title{A print method for \code{manyiv_est} objects}
\usage{
\method{summary}{manyiv_est}(object, ...)
}
\arguments{
\item{object}{A \code{manyiv_est} object from \code{manyiv}}

\item{...}{Unused}
}
\description{
A print method for \code{manyiv_est} objects
}
