% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ujive.R
\name{ujive}
\alias{ujive}
\title{Many IV regressions from Kolesár (2013)}
\usage{
ujive(data, y, exogenous, endogenous, instruments, cluster = NULL, ssc = FALSE)
}
\arguments{
\item{data}{Data.frame}

\item{y}{Character or formula. The outcome variable.}

\item{exogenous}{Formula. Outcome variable and exogenous control variables. Can include fixed effects in formula.}

\item{endogenous}{Formula. The endogenous variable of interest.}

\item{instruments}{Formula. Instruments for the endogenous variable. Can include fixed effects in formula.}

\item{cluster}{Character or formula. The cluster variable. For non-clustered robust standard errors, set to NULL.}

\item{ssc}{Logical. Should a small sample adjustment be made?}
}
\value{
An object of class \code{manyiv_est} that contains the results.
}
\description{
This routine is for a single endogenous regressor following Kolesár (2013).
The function produces a set of estimates using: OLS, two-stage least squares
estimator (TSLS), the limited information maximum likelihood (LIML),
the modified bias-correct 2SLS estimator (MB2SLS), the jackknife IV
estimator (JIVE), the unbiased jackknife IV estimator (UJIVE), and the
reverse two-stage least squares estimator (RTSLS). Details can be found in
https://www.princeton.edu/~mkolesar/papers/late_estimation.pdf.

For just UJIVE, see the dedicated function \code{ujive}.
}
