% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.jive_est}
\alias{summary.jive_est}
\title{A print method for \code{jive_est} objects}
\usage{
\method{summary}{jive_est}(x)
}
\arguments{
\item{x}{A \code{jive_est} object from \code{jive}/\code{ujive}/\code{ijive}}
}
\description{
A print method for \code{jive_est} objects
}
