% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ujive.R
\name{ujive}
\alias{ujive}
\title{Estimate unbiased jackknife IV regression from Kolesár (2013)}
\usage{
ujive(formula, data, cluster = NULL, ssc = FALSE, lo_cluster = FALSE)
}
\arguments{
\item{formula}{Formula. following the syntax of the \code{fixest} package. In short, \code{y ~ exo | exo_FEs | endo ~ instrument | instrument_FEs}.}

\item{data}{Data.frame}

\item{cluster}{Character or formula. The cluster variable. For non-clustered robust standard errors, set to NULL.}

\item{ssc}{Logical. Should a small sample adjustment be made? Default is \code{TRUE}.}

\item{lo_cluster}{Logical. Should leave-out cluster be used for estimation? Default is \code{TRUE} if \code{cluster} is not NULL.}
}
\value{
An object of class \code{jive_est} that contains the results. It contains the following information:
\item{beta}{Coefficient on the endogenous variable}
\item{se}{Standard error for the coefficient on the endogenous variable}
\item{F}{First-stage F statistic}
\item{Omega}{Estimate of the covariance matrix of reduced-form errors}
\item{Xi}{Estimate of the covariance matrix of the reduced form coefficients. See Kolesar (2012)}
\item{Sargan}{Sargan (1958) test for overidentifying restrictions. This is an empty list if only a single instrument is used.}
\item{CD}{Cragg and Donald (1993, 1997) test for overidentifying restrictions. This is an empty list if only a single instrument is used.}
\item{clustered}{Logical variable indicating whether standard errors are clustered.}
\item{n_clustered}{The number of clusters.}
\item{n}{The number of observations used.}
\item{n_instruments}{The number of non-collinear instruments included.}
\item{n_covariates}{The number of non-collinear covariates included.}
}
\description{
This routine is for a single endogenous regressor following Kolesár (2013).
The function produces a set of estimates using: OLS, two-stage least squares
estimator (TSLS), the limited information maximum likelihood (LIML),
the modified bias-correct 2SLS estimator (MB2SLS), the jackknife IV
estimator (JIVE), the unbiased jackknife IV estimator (UJIVE), and the
reverse two-stage least squares estimator (RTSLS). Details can be found in
https://www.princeton.edu/~mkolesar/papers/late_estimation.pdf.

For just UJIVE, see the dedicated function \code{ujive}.
}
